requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
dicttoxml==1.7.4
Faker==4.0.2
Flask==0.12.2
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
flask-restplus==0.13.0
Flask-SocketIO==2.9.6
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.15.7
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==3.12
six==1.14.0
Werkzeug==0.14.1
zipp==3.1.0
